buffer: rename `buffer` directory to `buffers`
