+++ src/core/buffers/active_period_emitter.ts
