+++ src/core/buffers/adaptation/index.ts
