+++ src/core/buffers/events_generators.ts
