+++ src/core/buffers/period/create_fake_buffer.ts
