+++ src/core/buffers/period/index.ts
