+++ src/core/buffers/period_buffer_manager.ts
