+++ src/core/buffers/representation/force_garbage_collection.ts
