+++ src/core/buffers/representation/get_wanted_range.ts
