+++ src/core/buffers/representation/index.ts
