+++ src/core/buffers/segment_bookkeeper.ts
