+++ src/core/init/types.ts
@@ -18,7 +18,7 @@ import { BehaviorSubject } from "rxjs";
 import { ICustomError } from "../../errors";
 import Manifest from "../../manifest";
 import ABRManager from "../abr";
-import { IRepresentationChangeEvent } from "../buffer/types";
+import { IRepresentationChangeEvent } from "../buffers/types";
 import { IStallingItem } from "./get_stalled_events";
 
 // Object emitted when the clock ticks
